Move unified input into its own file
